refactor(Toggle): switch to toggle
