+++ package.json
@@ -190,7 +190,7 @@
   },
   "dependencies": {
     "@babel/runtime": "7.18.9",
-    "@scaleway/ui": "0.178.3",
+    "@scaleway/ui": "0.183.3",
     "final-form": "4.20.7",
     "final-form-arrays": "3.0.2",
     "final-form-focus": "1.1.2",
