+++ src/index.ts
@@ -7,7 +7,7 @@ export { default as SelectNumberField } from './components/SelectNumberField'
 export { default as SelectableCardField } from './components/SelectableCardField'
 export { default as Submit } from './components/Submit'
 export { default as SubmitErrorAlert } from './components/SubmitErrorAlert'
-export { default as SwitchField } from './components/SwitchField'
+export { default as ToggleField } from './components/ToggleField'
 export { default as TagsField } from './components/TagsField'
 export { default as TextBoxField } from './components/TextBoxField'
 export { default as DateField } from './components/DateField'
